fix: minor refactors
